1. Renamed the state variables:
I replaced [isOpen, setIsOpen] with [isMenuOpen, setIsMenuOpen] to make it clearer that this state is responsible for opening and closing the menu.

2. Fixed the submission handler:

I removed the handler from the button and added it only to the <form> tag with the submit event. Now everything is working correctly, without duplication.

3. Optimized the effect:

Added a check so that handlers are not hung if the form is closed. Now the effect only works when it's really needed.

4. Brought the form in line with the Figma layout:

Added the title, text, and correctly positioned the inputs.

I have improved the form by adding the ability to edit the background color, which was previously missing.
